pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "sax", email = "s.apostolico@gmail.com"},
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Private"}

dependencies = [
    "django>=4.2.5",
    "django-admin-extra-buttons>=1.5.6",
    "django-adminactions>=2.1.0",
    "django-adminfilters>=2.2.0",
    "django-concurrency>=2.4",
    "django-cors-headers>=4.2.0",
    "django-debug-toolbar>=4.2.0",
    "django-flags>=5.0.13",
    "django-environ>=0.11.2",
    "django-smart-admin>=2.6.0",
    "django-tailwind>=3.6.0",
    "psycopg2-binary>=2.9.9",
    "social-auth-app-django>=5.3.0",
    "unicef-security>=1.4.4",
    "django-regex>=0.5.0",
    "djangorestframework>=3.14.0",
    "pillow>=10.0.0",
    "docutils>=0.20.1",
    "swapper>=1.3.0",
    "django-silk>=5.0.4",
    "django-push-notifications>=3.0.0",
    "htmlmin>=0.1.12",
    "django-bitfield>=2.2.0",
    "sentry-sdk>=1.31.0",
    "django-advanced-filters>=2.0.0",
    "django-celery-beat>=2.5.0",
    "django-mptt>=0.15.0",
    "natural-keys>=2.1.0",
    "tablib>=3.5.0",
    "django-import-export>=3.3.1",
    "redis>=5.0.1",
    "django-taggit>=4.0.0",
]
[tool.pdm.scripts]
lint = "pre-commit run --all-files"

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.2",
    "django-webtest>=1.9.10",
    "responses>=0.23.3",
    "pytest-django>=4.5.2",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
]
code = [
    "flake8>=6.1.0",
    "pdbpp>=0.10.3",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "black>=23.9.1",
    "pre-commit>=3.4.0",
    "types-freezegun>=1.1.10",
    "types-requests>=2.31.0.2",
    "types-python-dateutil>=2.8.19.14",
    "types-pytz>=2023.3.0.1",
    "types-redis>=4.6.0.5",
    "django-stubs>=4.2.4",
    "django-stubs-ext>=4.2.2",
    "types-setuptools>=68.2.0.0",
]


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | snapshots
)/
'''